feat(Gate): add 'IsAlternativeHeadquarterFor' relation
