+++ bpdm-gate-api/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/api/model/RelationType.kt
@@ -20,5 +20,6 @@
 package org.eclipse.tractusx.bpdm.gate.api.model
 
 enum class RelationType {
-    IsManagedBy
+    IsManagedBy,
+    IsAlternativeHeadquarterFor
 }
