+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/auth/AuthInputConsumerIT.kt
@@ -28,6 +28,7 @@ import org.eclipse.tractusx.bpdm.test.util.AuthAssertionHelper
 import org.eclipse.tractusx.bpdm.test.util.AuthExpectationType
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.ContextConfiguration
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = [Application::class])
@@ -36,6 +37,7 @@ import org.springframework.test.context.ContextConfiguration
     KeyCloakInitializer::class,
     AuthInputConsumerIT.SelfClientAsInputConsumerInitializer::class
 ])
+@ActiveProfiles("test")
 class AuthInputConsumerIT @Autowired constructor(
     gateClient: GateClient,
     authAssertionHelper: AuthAssertionHelper
