+++ bpdm-gate/src/test/resources/application-test-no-auth.yml
@@ -8,3 +8,7 @@ bpdm:
       security-enabled: false
     orchestrator:
       security-enabled: false
+spring:
+  datasource:
+    hikari:
+      maximum-pool-size: 2
